# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      security: [{ bearerAuth: [] }]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: {type: string, minLength: 1, maxLength: 255}
        size: {type: integer, minimum: 1, maximum: 10485760}
        contentType: {type: string, enum: [image/jpeg, image/png, image/webp]}
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: {type: string, format: uri}
        objectPath: {type: string}
        metadata: {$ref: "#/components/schemas/UploadUrlRequest"}
...